review suggestions from bls base
