+++ mithril-stm/src/bls_multi_signature/proof_of_possession.rs
@@ -1,6 +1,8 @@
-use crate::bls_multi_sig::signing_key::SigningKey;
-use crate::bls_multi_sig::unsafe_helpers::{compress_p1, scalar_to_pk_in_g1, uncompress_p1};
-use crate::bls_multi_sig::POP;
+use crate::bls_multi_signature::helper::unsafe_helpers::{
+    compress_p1, scalar_to_pk_in_g1, uncompress_p1,
+};
+use crate::bls_multi_signature::signing_key::SigningKey;
+use crate::bls_multi_signature::POP;
 use crate::error::{blst_err_to_mithril, MultiSignatureError};
 use blst::blst_p1;
 use blst::min_sig::Signature as BlstSig;
