+++ mithril-stm/src/bls_multi_signature/signing_key.rs
@@ -1,4 +1,4 @@
-use crate::bls_multi_sig::signature::Signature;
+use crate::bls_multi_signature::signature::Signature;
 use crate::error::{blst_err_to_mithril, MultiSignatureError};
 use blst::min_sig::SecretKey as BlstSk;
 use rand_core::{CryptoRng, RngCore};
