+++ mithril-stm/src/lib.rs
@@ -16,9 +16,9 @@ pub use crate::error::{
 };
 
 #[cfg(feature = "benchmark-internals")]
-pub mod bls_multi_sig;
+pub mod bls_multi_signature;
 
 #[cfg(not(feature = "benchmark-internals"))]
-pub mod bls_multi_sig;
+pub mod bls_multi_signature;
 
 pub mod participant;
