+++ mithril-stm/src/stm.rs
@@ -107,7 +107,7 @@
 //! # }
 //! ```
 
-use crate::bls_multi_sig::{Signature, VerificationKey};
+use crate::bls_multi_signature::{Signature, SigningKey, VerificationKey, VerificationKeyPoP};
 use crate::eligibility_check::ev_lt_phi;
 use crate::error::{
     AggregationError, CoreVerifierError, RegisterError, StmAggregateSignatureError,
